feat: update type hoc
